# openapi.yaml
openapi: "3.1.0"
info:
  title: Malluland API
  version: "1.0.0"
  description: Official API for Malluland backend services.
servers:
  - url: http://localhost:8787
    description: Local server
components:
  securitySchemes:
    XUserId:
      type: apiKey
      in: header
      name: x-user-id
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /users/settings:
    get:
      summary: Get current user's settings
      description: Returns the settings for the authenticated user.
      responses:
        '200':
          description: Settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    nullable: true
                    properties:
                      userId: { type: string, format: uuid }
                      chatAudience: { type: string, enum: [anyone, men, women, others], nullable: true }
                      pushEnabled: { type: boolean, nullable: true }
        '401': { description: Unauthorized }
    put:
      summary: Upsert current user's settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                chatAudience:
                  type: string
                  enum: [anyone, men, women, others]
                  nullable: true
                pushEnabled:
                  type: boolean
                  nullable: true
      responses:
        '200':
          description: Updated settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      userId: { type: string, format: uuid }
                      chatAudience: { type: string, enum: [anyone, men, women, others], nullable: true }
                      pushEnabled: { type: boolean, nullable: true }
        '401': { description: Unauthorized }
  /users/location:
    get:
      summary: Get current user's location
      description: Returns the saved location for the authenticated user.
      responses:
        '200':
          description: Location
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    nullable: true
                    properties:
                      userId: { type: string, format: uuid }
                      lat: { type: number, nullable: true }
                      lng: { type: number, nullable: true }
                      closestAirportCode: { type: string, nullable: true }
                      createdAt: { type: string, format: date-time, nullable: true }
                      updatedAt: { type: string, format: date-time, nullable: true }
        '401': { description: Unauthorized }
    put:
      summary: Upsert current user's location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                lat: { type: number, nullable: true }
                lng: { type: number, nullable: true }
                closestAirportCode: { type: string, nullable: true }
      responses:
        '200':
          description: Updated location
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      userId: { type: string, format: uuid }
                      lat: { type: number, nullable: true }
                      lng: { type: number, nullable: true }
                      closestAirportCode: { type: string, nullable: true }
                      createdAt: { type: string, format: date-time, nullable: true }
                      updatedAt: { type: string, format: date-time, nullable: true }
        '401': { description: Unauthorized }
  /users/{id}:
    get:
      summary: Get user by id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: User }
        '404': { description: Not found }
    put:
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200': { description: Updated user }
  /users/{id}/photos:
    post:
      summary: Add photo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [originalUrl, imageType, position]
              properties:
                originalUrl: { type: string, format: uri }
                optimizedUrl: { type: string, format: uri, nullable: true }
                imageType: { type: string }
                position: { type: integer }
      responses:
        '201': { description: Created }
    get:
      summary: List photos
      responses:
        '200': { description: Photos, content: { application/json: { schema: { type: array, items: { type: object } } } } }
  /users/{id}/selfies:
    post:
      summary: Add selfie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [selfieUrl]
              properties:
                selfieUrl: { type: string, format: uri }
      responses:
        '201': { description: Created }
    get:
      summary: List selfies
      responses:
        '200': { description: Selfies, content: { application/json: { schema: { type: array, items: { type: object } } } } }
  /users/{id}/interests:
    post:
      summary: Replace interests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [interestIds]
              properties:
                interestIds: { type: array, items: { type: integer } }
      responses:
        '201': { description: Updated }
    get:
      summary: List interests
      responses:
        '200': { description: Interests, content: { application/json: { schema: { type: array, items: { type: object } } } } }
  /users/{id}/traits:
    post:
      summary: Replace traits
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [traitIds]
              properties:
                traitIds: { type: array, items: { type: integer } }
      responses:
        '201': { description: Updated }
    get:
      summary: List traits
      responses:
        '200': { description: Traits, content: { application/json: { schema: { type: array, items: { type: object } } } } }
  /users/{id}/favorite-actors:
    post:
      summary: Replace favorite actors
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [actorIds]
              properties:
                actorIds: { type: array, items: { type: integer } }
      responses:
        '201': { description: Updated }
    get:
      summary: List favorite actors
      responses:
        '200': { description: Favorites, content: { application/json: { schema: { type: array, items: { type: object } } } } }
  /users/{id}/favorite-actresses:
    post:
      summary: Replace favorite actresses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [actressIds]
              properties:
                actressIds: { type: array, items: { type: integer } }
      responses:
        '201': { description: Updated }
    get:
      summary: List favorite actresses
      responses:
        '200': { description: Favorites, content: { application/json: { schema: { type: array, items: { type: object } } } } }
  /users/{id}/social-links:
    post:
      summary: Replace social links
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [links]
              properties:
                links:
                  type: array
                  items:
                    type: object
                    properties:
                      platform: { type: string }
                      handle: { type: string }
      responses:
        '201': { description: Updated }
    get:
      summary: List social links
      responses:
        '200': { description: Social links, content: { application/json: { schema: { type: array, items: { type: object } } } } }
  /users/user-favorites:
    get:
      summary: List current user's favorite text entries
      parameters:
        - name: category
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Favorites text
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        userId: { type: string, format: uuid }
                        category: { type: string }
                        textValue: { type: string }
                        position: { type: integer }
    put:
      summary: Replace all favorite text for a category (max 5)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category, values]
              properties:
                category: { type: string }
                values:
                  type: array
                  maxItems: 5
                  items: { type: string, maxLength: 150 }
      responses:
        '200':
          description: Updated favorites
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        userId: { type: string, format: uuid }
                        category: { type: string }
                        textValue: { type: string }
                        position: { type: integer }
    post:
      summary: Replace all favorite text for a category (same as PUT)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category, values]
              properties:
                category: { type: string }
                values:
                  type: array
                  maxItems: 5
                  items: { type: string, maxLength: 150 }
      responses:
        '200':
          description: Updated favorites
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        userId: { type: string, format: uuid }
                        category: { type: string }
                        textValue: { type: string }
                        position: { type: integer }
  /users/user-favorites/add:
    post:
      summary: Add a single favorite text for a category (up to 5)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category, value]
              properties:
                category: { type: string }
                value: { type: string, maxLength: 150 }
      responses:
        '200':
          description: Added favorite
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      userId: { type: string, format: uuid }
                      category: { type: string }
                      textValue: { type: string }
                      position: { type: integer }
...